updating to integer index from long long

+++ src/solvers/pearson.cpp
@@ -6,9 +6,9 @@
 #include <array>
 #include "pearson.h"
 #include "../utils/moments.h"
-constexpr long long klen  = 256; 
+constexpr int klen  = 256; 
 
-static void dfdg_init(const double* __restrict__ ts, const double* __restrict__ mu, double* __restrict__ df, double* __restrict__ dg, long long len, long long sublen){
+static void dfdg_init(const double* __restrict__ ts, const double* __restrict__ mu, double* __restrict__ df, double* __restrict__ dg, int len, int sublen){
    df[0] = 0;
    dg[0] = 0;
    for(int i = 0; i < len - sublen; i++){
@@ -20,62 +20,62 @@ static void dfdg_init(const double* __restrict__ ts, const double* __restrict__ 
 static inline void pauto_pearson_kern(
    double*       __restrict__  cov,
    double*       __restrict__  mp,
-   long long*    __restrict__ mpi,
+   int*    __restrict__ mpi,
    const double* __restrict__ df,
    const double* __restrict__ dg,
    const double* __restrict__ invn,
-   const long long ofr,
-   const long long ofc,
-   const long long iters)
+   const int ofr,
+   const int ofc,
+   const int iters)
 {
    df   = (const double*)__builtin_assume_aligned(df, prefalign);
    dg   = (const double*)__builtin_assume_aligned(dg, prefalign);
    mp   = (double*) __builtin_assume_aligned(mp, prefalign);
-   mpi  = (long long*) __builtin_assume_aligned(mpi, prefalign);
+   mpi  = (int*) __builtin_assume_aligned(mpi, prefalign);
    invn = (const double*)__builtin_assume_aligned(invn, prefalign);
    cov  = (double*)__builtin_assume_aligned(cov, prefalign);
-   for(long long r = 0; r < iters; r++){
+   for(int r = 0; r < iters; r++){
       std::array<double, klen> cr;
       if(r != 0){
-         for(long long d = 0; d < klen; d++){ 
+         for(int d = 0; d < klen; d++){ 
             cov[d] += df[r] * dg[r + d + ofc];
             cov[d] += df[r + d + ofc] * dg[r];
          }
       }
-      for(long long d = 0; d < klen; d++){
+      for(int d = 0; d < klen; d++){
          cr[d] = cov[d] * invn[r] * invn[r + d + ofc];
          if(cov[d] * invn[r] * invn[r + d + ofc] > mp[r + d + ofc]){
             mp[r + d + ofc] = cov[d] * invn[r] * invn[r + d + ofc];
             mpi[r + d + ofc] = r + ofr;
          }
       }
-      std::array<long long, klen/2> ci;
-      for(long long i = 0; i < 128; i++){
+      std::array<int, klen/2> ci;
+      for(int i = 0; i < 128; i++){
          ci[i] = cr[i] > cr[i + 128] ? i : i + 128;
          cr[i] = cr[i] > cr[i + 128] ? cr[i] : cr[i + 128];
       }
-      for(long long i = 0; i < 64; i++){ // gcc can generally simplify conditional writes using if statements to masked writes, but in this case
+      for(int i = 0; i < 64; i++){ // gcc can generally simplify conditional writes using if statements to masked writes, but in this case
                                          // it generates all scalar code. For now I rewrote it using ternary operators. I'll file a bug report later.
          ci[i] = cr[i] > cr[i + 64] ? ci[i] : ci[i + 64];
          cr[i] = cr[i] > cr[i + 64] ? cr[i] : cr[i + 64];
       }
-      for(long long i = 0; i < 32; i++){
+      for(int i = 0; i < 32; i++){
          ci[i] = cr[i] > cr[i + 32] ? ci[i] : ci[i + 32];
          cr[i] = cr[i] > cr[i + 32] ? cr[i] : cr[i + 32];
       } 
-      for(long long i = 0; i < 16; i++){
+      for(int i = 0; i < 16; i++){
          ci[i] = cr[i] > cr[i + 16] ? ci[i] : ci[i + 16];
          cr[i] = cr[i] > cr[i + 16] ? cr[i] : cr[i + 16];
       }
-      for(long long i = 0; i < 8; i++){
+      for(int i = 0; i < 8; i++){
          ci[i] = cr[i] > cr[i + 8] ? ci[i] : ci[i + 8];
          cr[i] = cr[i] > cr[i + 8] ? cr[i] : cr[i + 8];
       }
-      for(long long i = 0; i < 4; i++){
+      for(int i = 0; i < 4; i++){
          ci[i] = cr[i] > cr[i + 4] ? ci[i] : ci[i + 4];
          cr[i] = cr[i] > cr[i + 4] ? cr[i] : cr[i + 4];
       }
-      for(long long i = 0; i < 2; i++){
+      for(int i = 0; i < 2; i++){
          ci[i] = cr[i] > cr[i + 2] ? ci[i] : ci[i + 2];
          cr[i] = cr[i] > cr[i + 2] ? cr[i] : cr[i+2];
       }
@@ -89,18 +89,18 @@ static inline void pauto_pearson_kern(
 static inline void pauto_pearson_edge(
    double*       __restrict__ cov, 
    double*       __restrict__ mp,  
-   long long*    __restrict__ mpi, 
+   int*    __restrict__ mpi, 
    const double* __restrict__ df,  
    const double* __restrict__ dg, 
    const double* __restrict__ invn, 
-   const long long ofr, 
-   const long long ofd, 
-   const long long dlim,
-   const long long clim,
+   const int ofr, 
+   const int ofd, 
+   const int dlim,
+   const int clim,
    const bool init)
 {
-   for(long long r = ofr; r < clim - ofd; r++){
-      for(long long d = ofd; d < std::min(clim - r, dlim); d++){
+   for(int r = ofr; r < clim - ofd; r++){
+      for(int d = ofd; d < std::min(clim - r, dlim); d++){
          if(!init || r != ofr){
             cov[d - ofd] += df[r] * dg[r + d];
             cov[d - ofd] += df[r + d] * dg[r];
@@ -118,14 +118,14 @@ static inline void pauto_pearson_edge(
    }
 }
 
-int partialauto(bufd& ts, bufd& mp, bufi& mpi, long long minlag, long long sublen){
+int partialauto(bufd& ts, bufd& mp, bufi& mpi, int minlag, int sublen){
    if(!(ts.valid() && mp.valid() && mpi.valid())){
       return errs::bad_input;  // Todo: Build a real set of error checking functions 
    }
    const int qstride = paddedlen(sublen, prefalign);
-   const long long mlen = ts.len - sublen + 1;
-   const long long tlen = std::max(static_cast<long long>(2 << 14), 4 * sublen - (4 * sublen) % klen);        
-   const long long tilesperdim = (mlen - minlag)/tlen + ((mlen - minlag) % tlen ? 1 : 0);
+   const int mlen = ts.len - sublen + 1;
+   const int tlen = std::max(static_cast<int>(2 << 14), 4 * sublen - (4 * sublen) % klen);        
+   const int tilesperdim = (mlen - minlag)/tlen + ((mlen - minlag) % tlen ? 1 : 0);
    bufd mu(mlen); bufd invn(mlen); bufd df(mlen);  
    bufd dg(mlen); bufd cov(mlen);  bufd q(tilesperdim * qstride);
    if(!(mu.valid() && df.valid() && dg.valid() && invn.valid())){
@@ -137,19 +137,19 @@ int partialauto(bufd& ts, bufd& mp, bufi& mpi, long long minlag, long long suble
 
    dfdg_init(ts(), mu(), df(), dg(), ts.len, sublen);
    #pragma omp parallel for
-   for(long long i = 0; i < tilesperdim; i++){
+   for(int i = 0; i < tilesperdim; i++){
       center_query(ts(i * tlen), mu(i * tlen), q(i * qstride), sublen); 
    }
-   for(long long diag = 0; diag < tilesperdim; diag++){
+   for(int diag = 0; diag < tilesperdim; diag++){
       #pragma omp parallel for 
-      for(long long ofst = 0; ofst < tilesperdim - diag; ofst++){
-         const long long di = diag * tlen + minlag;
-         const long long ofi = ofst * tlen;
-         const long long dlim = std::min(di + tlen, mlen - ofi);
+      for(int ofst = 0; ofst < tilesperdim - diag; ofst++){
+         const int di = diag * tlen + minlag;
+         const int ofi = ofst * tlen;
+         const int dlim = std::min(di + tlen, mlen - ofi);
          autocov(ts(di + ofi), mu(di + ofi), q(ofst * qstride), cov(ofi), dlim - di, sublen);
-         for(long long d = di; d < dlim; d += klen){
+         for(int d = di; d < dlim; d += klen){
             if(d + klen <= dlim){
-               const long long ral = std::min(tlen, mlen - d - ofi - klen + 1); 
+               const int ral = std::min(tlen, mlen - d - ofi - klen + 1); 
                pauto_pearson_kern(cov(ofi + d - di), mp(ofi), mpi(ofi), df(ofi), dg(ofi), invn(ofi), ofi, d, ral);
                if(ral < tlen){
                   pauto_pearson_edge(cov(ofi + d - di), mp(), mpi(), df(), dg(), invn(), ofi + ral, d, d + klen, mlen, false);
